requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
Pillow==10.1.0